Add a __restore_rt arm64 trampoline
The unwind info here doesn't actually work:

It restores the LR, and uses that as the return address column, but the
LR and PC are different for a signal frame. There doesn't appear to be
a DWARF register number reserved for the PC[1], so maybe there's no way
to fix the FDE, and the unwinder need to look for the sigreturn
instructions instead.

[1] ihi0057, DWARF for the Arm 64-bit Architecture

Change-Id: Ibeaa27d8d258bd6ed442715b839fa3676cd99c55

---

The FDE uses DW_CFA_def_cfa_expression and DW_CFA_expression CFI
instructions to accommodate the variable offset between the restored
SP (i.e. the CFA) and the fields of the sigcontext holding saved
register values.

This FDE only restores GPRs (x0..x31) and the CFA/SP. The arm64 ABI
(IHI0055C_beta_aapcs64) also specifies that the lower 64 bits of v8-v15
should be preserved. Currently, this FDE doesn't attempt to restore
them.

With this FDE, LLVM's libunwind should be able to unwind through a
signal frame.

Test: bionic-unit-tests
Test: step through frames in lldb/gdb (lldb shows a missing frame)
Test: dump _Unwind_Backtrace callbacks (also shows a missing frame)
Change-Id: I61c5b86ea59c183566816e902517451dc5d25913
Signed-off-by: Albert I <kras@raphielgang.org>